Move styles to /style
